// src/Anfx.Pasivos.Frontend/src/api/models/carteraReporteDtoPagedResultDtoApiResponseDto.ts
/**
 * Anfx.Pasivos.ApiService
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CarteraReporteDtoPagedResultDto } from './carteraReporteDtoPagedResultDto';


export interface CarteraReporteDtoPagedResultDtoApiResponseDto { 
    success?: boolean;
    message?: string | null;
    data?: CarteraReporteDtoPagedResultDto;
    errors?: Array<string> | null;
    statusCode?: number;
    timestamp?: string;
    traceId?: string | null;
}